feat: 게시글 좋아요 삭제 테스트 및 구조 변경
