+++ backend/src/test/java/com/backjoongwon/cvi/post/application/PostServiceTest.java
@@ -3,6 +3,7 @@ package com.backjoongwon.cvi.post.application;
 import com.backjoongwon.cvi.auth.domain.authorization.SocialProvider;
 import com.backjoongwon.cvi.common.exception.InvalidOperationException;
 import com.backjoongwon.cvi.common.exception.NotFoundException;
+import com.backjoongwon.cvi.post.dto.LikeResponse;
 import com.backjoongwon.cvi.post.domain.Post;
 import com.backjoongwon.cvi.post.domain.PostRepository;
 import com.backjoongwon.cvi.post.domain.VaccinationType;
@@ -23,6 +24,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Stream;
@@ -44,9 +47,15 @@ class PostServiceTest {
 
     @Autowired
     private PostService postService;
+
+    @PersistenceContext
+    private EntityManager em;
+
     private User user;
+    private User anotherUser;
     private Post post;
     private PostRequest postRequest;
+    private LikeResponse likeResponse;
 
     @BeforeEach
     void init() {
@@ -57,14 +66,23 @@ class PostServiceTest {
                 .socialProvider(SocialProvider.NAVER)
                 .build();
         userRepository.save(user);
+        anotherUser = User.builder()
+                .nickname("다른_유저")
+                .ageRange(AgeRange.TWENTIES)
+                .profileUrl("")
+                .socialProvider(SocialProvider.KAKAO)
+                .build();
+        userRepository.save(anotherUser);
         post = Post.builder()
                 .content("Test Content111")
                 .vaccinationType(VaccinationType.ASTRAZENECA)
                 .user(user)
                 .createdAt(LocalDateTime.now())
                 .build();
-        postRequest = new PostRequest("Test Content222", VaccinationType.PFIZER);
         postRepository.save(post);
+        likeResponse = postService.createLike(post.getId(), RequestUser.of(user.getId()));
+        postRequest = new PostRequest("Test Content222", VaccinationType.PFIZER);
+        resetEntityManager();
     }
 
     @DisplayName("게시글 생성 - 성공")
@@ -231,13 +249,71 @@ class PostServiceTest {
                 .user(user)
                 .createdAt(LocalDateTime.now())
                 .build();
-        postRepository.save(post);
 
+        postRepository.save(post);
         //when
         List<PostResponse> postResponses = postService.findByVaccineType(vaccinationType);
         //then
         assertThat(postResponses).filteredOn(
                 response -> response.getVaccinationType().equals(vaccinationType)
         );
+    }
+
+    @DisplayName("게시글 좋아요 생성 - 성공")
+    @Test
+    void createLike() {
+        //given
+        //when
+        LikeResponse like = postService.createLike(post.getId(), RequestUser.of(anotherUser.getId()));
+
+        resetEntityManager();
+        //then
+        Post post = getPost();
+        assertThat(post.getLikesCount()).isEqualTo(2);
+    }
+
+    @DisplayName("게시글 좋아요 생성 - 실패 - 게시글이 없는 경우")
+    @Test
+    void createLikeFailureWhenPostNotExists() {
+        //given
+        //when
+        //then
+        assertThatThrownBy(() -> postService.createLike(post.getId() + 1L, RequestUser.of(anotherUser.getId())))
+                .isInstanceOf(NotFoundException.class);
+    }
+
+    @DisplayName("게시글 좋아요 생성 - 실패 - 동일한 유저가 이미 좋아요를 누른 경우")
+    @Test
+    void createLikeFailureWhenAlreadyCreatedBySameUser() {
+        //given
+        //when
+        //then
+        assertThatThrownBy(() -> postService.createLike(post.getId(), RequestUser.of(user.getId())))
+                .isInstanceOf(InvalidOperationException.class);
+    }
+
+    @DisplayName("게시글 좋아요 삭제 - 성공")
+    @Test
+    void deleteLike() {
+        //given
+        RequestUser requestUser = RequestUser.of(user.getId());
+        //when
+        postService.deleteLike(post.getId(), likeResponse.getId(), requestUser);
+        resetEntityManager();
+        //then
+        Post actualPost = postRepository.findWithLikesById(this.post.getId())
+                .orElseThrow(() -> new NotFoundException("해당 id의 게시글이 존재하지 않습니다."));
+        assertThat(actualPost.getLikes()).isEmpty();
+    }
+
+    private void resetEntityManager() {
+        em.flush();
+        em.clear();
+        em.close();
+    }
+
+    private Post getPost() {
+        return postRepository.findWithLikesById(post.getId())
+                .orElseThrow(() -> new NotFoundException("해당 id의 게시글이 존재하지 않습니다."));
     }
 }
